# public/swagger/v1/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Meta": {
				"properties": {
					"publisher": {
						"type": "string"
					},
					"licence": {
						"type": "string"
					},
					"documentation": {
						"type": "string"
					},
					"version": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					},
					"hasFormat": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"publisher",
					"licence",
					"documentation",
					"version",
					"comment",
					"hasFormat",
					"limit"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Allergen": {
				"properties": {
					"id": {
						"type": "string"
					},
					"altLabel": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"narrower": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"notation": {
						"type": "string"
					},
					"riskStatement": {
						"type": "string"
					},
					"broader": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"label",
					"notation",
					"riskStatement"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Problem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"allergen": {
						"items": {
							"$ref": "#/components/schemas/Allergen"
						},
						"type": "array"
					},
					"riskStatement": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"riskStatement"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BatchDescription": {
				"properties": {
					"id": {
						"type": "string"
					},
					"batchCode": {
						"type": "string"
					},
					"bestBeforeDate": {
						"type": "string",
						"format": "date-time"
					},
					"bestBeforeDescription": {
						"type": "string"
					},
					"useByDescription": {
						"type": "string"
					},
					"lotNumber": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductDetail": {
				"properties": {
					"id": {
						"type": "string"
					},
					"batchDescription": {
						"items": {
							"$ref": "#/components/schemas/BatchDescription"
						},
						"type": "array"
					},
					"packSizeDescription": {
						"type": "string"
					},
					"productName": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"batchDescription",
					"productName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Business": {
				"properties": {
					"commonName": {
						"type": "string"
					}
				},
				"required": [
					"commonName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Label": {
				"enum": [
					"Published"
				],
				"type": "string"
			},
			"Status": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"$ref": "#/components/schemas/Label"
					}
				},
				"required": [
					"id",
					"label"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemFull": {
				"properties": {
					"id": {
						"type": "string"
					},
					"smStext": {
						"type": "string"
					},
					"actionTaken": {
						"type": "string"
					},
					"alertURL": {
						"type": "string"
					},
					"consumerAdvice": {
						"type": "string"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"description": {
						"type": "string"
					},
					"modified": {
						"type": "string",
						"format": "date-time"
					},
					"notation": {
						"type": "string"
					},
					"problem": {
						"items": {
							"$ref": "#/components/schemas/Problem"
						},
						"type": "array"
					},
					"productDetails": {
						"items": {
							"$ref": "#/components/schemas/ProductDetail"
						},
						"type": "array"
					},
					"relatedMedia": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"reportingBusiness": {
						"$ref": "#/components/schemas/Business"
					},
					"shortTitle": {
						"type": "string"
					},
					"shortURL": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"title": {
						"type": "string"
					},
					"twitterText": {
						"type": "string"
					},
					"type": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"otherBusiness": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/Business"
								},
								"type": "array"
							},
							{
								"$ref": "#/components/schemas/Business"
							}
						]
					}
				},
				"required": [
					"id",
					"smStext",
					"actionTaken",
					"alertURL",
					"consumerAdvice",
					"created",
					"description",
					"modified",
					"notation",
					"problem",
					"productDetails",
					"reportingBusiness",
					"shortTitle",
					"shortURL",
					"status",
					"title",
					"twitterText",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IListAlertGetFullResponse": {
				"properties": {
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/ItemFull"
						},
						"type": "array"
					}
				},
				"required": [
					"meta",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReportingBusiness": {
				"properties": {
					"commonName": {
						"type": "string"
					}
				},
				"required": [
					"commonName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemDefault": {
				"properties": {
					"id": {
						"type": "string"
					},
					"alertURL": {
						"type": "string"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"modified": {
						"type": "string",
						"format": "date-time"
					},
					"notation": {
						"type": "string"
					},
					"problem": {
						"items": {
							"$ref": "#/components/schemas/Problem"
						},
						"type": "array"
					},
					"productDetails": {
						"items": {
							"$ref": "#/components/schemas/ProductDetail"
						},
						"type": "array"
					},
					"reportingBusiness": {
						"$ref": "#/components/schemas/ReportingBusiness"
					},
					"shortTitle": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"title": {
						"type": "string"
					},
					"type": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"alertURL",
					"created",
					"modified",
					"notation",
					"problem",
					"productDetails",
					"reportingBusiness",
					"shortTitle",
					"status",
					"title",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IListAlertGetDefaultResponse": {
				"properties": {
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/ItemDefault"
						},
						"type": "array"
					}
				},
				"required": [
					"meta",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_VIEW": {
				"description": "*************************      Enum Constants      ****************************************",
				"enum": [
					"default",
					"full"
				],
				"type": "string"
			},
			"AlertURL": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RelatedMedia": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemIndividual": {
				"properties": {
					"id": {
						"type": "string"
					},
					"smStext": {
						"type": "string"
					},
					"actionTaken": {
						"type": "string"
					},
					"alertURL": {
						"$ref": "#/components/schemas/AlertURL"
					},
					"consumerAdvice": {
						"type": "string"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"description": {
						"type": "string"
					},
					"modified": {
						"type": "string",
						"format": "date-time"
					},
					"notation": {
						"type": "string"
					},
					"problem": {
						"items": {
							"$ref": "#/components/schemas/Problem"
						},
						"type": "array"
					},
					"productDetails": {
						"items": {
							"$ref": "#/components/schemas/ProductDetail"
						},
						"type": "array"
					},
					"relatedMedia": {
						"items": {
							"$ref": "#/components/schemas/RelatedMedia"
						},
						"type": "array"
					},
					"reportingBusiness": {
						"$ref": "#/components/schemas/ReportingBusiness"
					},
					"shortTitle": {
						"type": "string"
					},
					"shortURL": {
						"$ref": "#/components/schemas/AlertURL"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"title": {
						"type": "string"
					},
					"twitterText": {
						"type": "string"
					},
					"type": {
						"items": {
							"$ref": "#/components/schemas/AlertURL"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"smStext",
					"actionTaken",
					"alertURL",
					"consumerAdvice",
					"created",
					"description",
					"modified",
					"notation",
					"problem",
					"productDetails",
					"relatedMedia",
					"reportingBusiness",
					"shortTitle",
					"shortURL",
					"status",
					"title",
					"twitterText",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IFoodAlertIndividualAlertGetResponse": {
				"properties": {
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/ItemIndividual"
						},
						"type": "array"
					}
				},
				"required": [
					"meta",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemAllergen": {
				"properties": {
					"id": {
						"type": "string"
					},
					"broader": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"notation": {
						"type": "string"
					},
					"riskStatement": {
						"type": "string"
					},
					"altLabel": {
						"type": "string"
					},
					"narrower": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"label",
					"notation",
					"riskStatement"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IFoodAlertListCodeAllergensGetResponse": {
				"properties": {
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/ItemAllergen"
						},
						"type": "array"
					}
				},
				"required": [
					"meta",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pathogen": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"type": "string"
							}
						]
					},
					"notation": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"label",
					"notation"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemPathogen": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"type": "string"
							}
						]
					},
					"notation": {
						"type": "string"
					},
					"pathogen": {
						"$ref": "#/components/schemas/Pathogen"
					},
					"riskStatement": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"id",
					"label",
					"notation",
					"pathogen",
					"riskStatement"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IFoodAlertPathogenRiskGetResponse": {
				"properties": {
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/ItemPathogen"
						},
						"type": "array"
					}
				},
				"required": [
					"meta",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemHazardCatagories": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"notation": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"label",
					"notation"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IFoodAlertHazardCatagoryGetResponse": {
				"properties": {
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/ItemHazardCatagories"
						},
						"type": "array"
					}
				},
				"required": [
					"meta",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IFoodAlertReasonAlertGetResponse": {
				"properties": {
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/ItemHazardCatagories"
						},
						"type": "array"
					}
				},
				"required": [
					"meta",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"name": "authorization",
				"in": "header",
				"bearerFormat": "JWT"
			},
			"apiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "X-API-KEY"
			}
		}
	},
	"info": {
		"title": "Food Standards Agency API",
		"version": "v1",
		"description": "The FSA Food Alerts API provides access to current and recent Food Alerts: Allergy Alerts (AA), Product Recall Information Notices (PRIN) and Food Alerts for Action (FAFA). It provides applications with the facility to list alerts matching some filter criterion, and to retrieve a description of an alert.",
		"contact": {
			"name": "Oraclo LTD",
			"email": "admin@orcalo.co.uk",
			"url": "https://orcalo.co.uk/"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/v1/foodalerts/id": {
			"get": {
				"operationId": "GetListAlerts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IListAlertGetFullResponse"
										},
										{
											"$ref": "#/components/schemas/IListAlertGetDefaultResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"meta": {
												"publisher": "string",
												"licence": "string",
												"documentation": "string",
												"version": "string",
												"comment": "string",
												"hasFormat": [
													"string"
												],
												"limit": 0
											},
											"items": [
												{
													"id": "string",
													"smStext": "string",
													"actionTaken": "string",
													"alertURL": "string",
													"consumerAdvice": "string",
													"created": "2021-09-13T13:22:53.195Z",
													"description": "string",
													"modified": "2021-09-13T13:22:53.195Z",
													"notation": "string",
													"problem": [
														{
															"id": "string",
															"allergen": [
																{
																	"id": "string",
																	"altLabel": "string",
																	"label": "string",
																	"narrower": [
																		"string"
																	],
																	"notation": "string",
																	"riskStatement": "string",
																	"broader": "string"
																}
															],
															"riskStatement": "string"
														}
													],
													"productDetails": [
														{
															"id": "string",
															"batchDescription": [
																{
																	"id": "string",
																	"batchCode": "string",
																	"bestBeforeDate": "2021-09-13T13:22:53.195Z",
																	"bestBeforeDescription": "string",
																	"useByDescription": "string",
																	"lotNumber": "string"
																}
															],
															"packSizeDescription": "string",
															"productName": "string"
														}
													],
													"relatedMedia": [
														"string"
													],
													"reportingBusiness": {
														"commonName": "string"
													},
													"shortTitle": "string",
													"shortURL": "string",
													"status": {
														"id": "string",
														"label": "Published"
													},
													"title": "string",
													"twitterText": "string",
													"type": [
														"string"
													],
													"otherBusiness": [
														{
															"commonName": "string"
														}
													]
												}
											]
										}
									},
									"Example 2": {
										"value": {
											"meta": {
												"publisher": "string",
												"licence": "string",
												"documentation": "string",
												"version": "string",
												"comment": "string",
												"hasFormat": [
													"string"
												],
												"limit": 0
											},
											"items": [
												{
													"id": "string",
													"actionTaken": "string",
													"alertURL": "string",
													"consumerAdvice": "string",
													"created": "2021-09-13T13:22:53.195Z",
													"description": "string",
													"modified": "2021-09-13T13:22:53.195Z",
													"notation": "string",
													"problem": [
														{
															"id": "string",
															"allergen": [
																{
																	"id": "string",
																	"altLabel": "string",
																	"label": "string",
																	"narrower": [
																		"string"
																	],
																	"notation": "string",
																	"riskStatement": "string",
																	"broader": "string"
																}
															],
															"riskStatement": "string"
														}
													],
													"productDetails": [
														{
															"id": "string",
															"batchDescription": [
																{
																	"id": "string",
																	"batchCode": "string",
																	"bestBeforeDate": "2021-09-13T13:22:53.195Z",
																	"bestBeforeDescription": "string",
																	"useByDescription": "string",
																	"lotNumber": "string"
																}
															],
															"packSizeDescription": "string",
															"productName": "string"
														}
													],
													"relatedMedia": [
														"string"
													],
													"reportingBusiness": {
														"commonName": "string"
													},
													"shortTitle": "string",
													"shortURL": "string",
													"status": {
														"id": "string",
														"label": "Published"
													},
													"title": "string",
													"twitterText": "string",
													"type": [
														"string"
													],
													"otherBusiness": [
														{
															"commonName": "string"
														}
													]
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"summary": "List Alerts <b>|</b> List alerts changed since some point in time <b>|</b> Search for alerts matching some search term",
				"tags": [
					"Food Alerts"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be greater than <b>0</b>",
						"in": "query",
						"name": "_limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Must be <b>\"default\"</b> or <b>\"full\"</b>",
						"in": "query",
						"name": "_view",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/_VIEW"
						}
					},
					{
						"description": "List alerts changed since some point in time <b>(Example: 2018-01-29T16:25:08+00:00)</b>",
						"in": "query",
						"name": "since",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Search for alerts matching some search term <b>(Example: tesco)</b>",
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/foodalerts/id/{id}": {
			"get": {
				"operationId": "GetIndividualAlert",
				"responses": {
					"200": {
						"description": "IFoodAlertIndividualAlertGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IFoodAlertIndividualAlertGetResponse"
								}
							}
						}
					}
				},
				"summary": "An individual alert",
				"tags": [
					"Food Alerts"
				],
				"security": [],
				"parameters": [
					{
						"description": "Example = FSA-AA-01-2018",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/foodalerts/def/allergens": {
			"get": {
				"operationId": "GetListCodeAllergens",
				"responses": {
					"200": {
						"description": "IFoodAlertListCodeAllergensGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IFoodAlertListCodeAllergensGetResponse"
								}
							}
						}
					}
				},
				"summary": "List of codes for allergens",
				"tags": [
					"Food Alerts"
				],
				"security": [],
				"parameters": [
					{
						"description": "Example = label",
						"in": "query",
						"name": "_sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/foodalerts/def/pathogen-risk": {
			"get": {
				"operationId": "GetListCodePathogenRisk",
				"responses": {
					"200": {
						"description": "IFoodAlertPathogenRiskGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IFoodAlertPathogenRiskGetResponse"
								}
							}
						}
					}
				},
				"summary": "List of codes for pathogen risks",
				"tags": [
					"Food Alerts"
				],
				"security": [],
				"parameters": [
					{
						"description": "Example = label",
						"in": "query",
						"name": "_sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/foodalerts/def/hazards": {
			"get": {
				"operationId": "GetListHazardsCatagories",
				"responses": {
					"200": {
						"description": "IFoodAlertHazardCatagoryGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IFoodAlertHazardCatagoryGetResponse"
								}
							}
						}
					}
				},
				"summary": "List of hazard categories",
				"tags": [
					"Food Alerts"
				],
				"security": [],
				"parameters": [
					{
						"description": "Example = label",
						"in": "query",
						"name": "_sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/foodalerts/def/reasons": {
			"get": {
				"operationId": "GetListReasonsAlert",
				"responses": {
					"200": {
						"description": "IFoodAlertReasonAlertGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IFoodAlertReasonAlertGetResponse"
								}
							}
						}
					}
				},
				"summary": "List of reasons for alert",
				"tags": [
					"Food Alerts"
				],
				"security": [],
				"parameters": [
					{
						"description": "Example = label",
						"in": "query",
						"name": "_sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}